refactor: move MultiSelect component, remove @expo/vector-icons to use lucide icons and changed selected element color to match the selection counting label
