+++ package.json
@@ -10,7 +10,6 @@
     "format:write": "prettier --write ."
   },
   "dependencies": {
-    "@expo/vector-icons": "^15.0.2",
     "@gluestack-ui/alert-dialog": "^0.1.38",
     "@gluestack-ui/button": "^1.0.14",
     "@gluestack-ui/form-control": "^0.1.19",
